pyproject.toml
[tool.poetry]
name = "searchcode"
version = "0.6.0"
description = "Simple, comprehensive code search."
authors = ["Ritchie Mwewa <rly0nheart@duck.com>"]
license = "GPLv3+"
readme = "README.md"
homepage = "https://searchcode.com"
repository = "https://github.com/rly0nheart/searchcode-sdk"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Natural Language :: English"
]


packages = [
    { include = "searchcode", from = "src" }
]
[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.2"
rich-click = "^1.8.8"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.2"
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
sc = "searchcode._cli.app:cli"
searchcode = "searchcode._cli.app:cli"